+++ app/Http/Requests/UpdateEvent.php
@@ -25,6 +25,7 @@ class UpdateEvent extends FormRequest
     {
         return [
             'is_online' => 'required|boolean',
+            'show_on_homepage' => 'required|boolean',
             'name' => 'bail|required:string',
             'event_type_id' => 'exists:event_types,id|required',
             'import_only' => 'required|boolean',