+++ app/Providers/AppServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace App\Providers;
 
+use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -12,7 +13,10 @@ class AppServiceProvider extends ServiceProvider
      * @return void
      */
     public function boot()
-    {//
+    {
+        if (env('DB_LOWER_STRING_LENGTH')) {
+            Schema::defaultStringLength(191);
+        }
     }
 
     /**